refactor : badge에 tailwind 속성 부여, containers 디렉토리로 이동
