+++ caffhheiene/src/constants/badge.tsx
@@ -20,90 +20,90 @@ import {
 
 export const BADGES = Object.freeze({
   css3: {
-    svg: <CSS3SVG width="20px" height="20px" />,
+    svg: <CSS3SVG className="h-4 w-4" />,
     color: '#1572B6',
     name: 'CSS3',
   },
   github: {
-    svg: <GithubSVG width="20px" height="20px" />,
+    svg: <GithubSVG className="h-4 w-4" />,
     color: '#000000',
     name: 'Github',
     link: 'https://github.com/SWARVY',
   },
   html5: {
-    svg: <Html5SVG width="20px" height="20px" />,
+    svg: <Html5SVG className="h-4 w-4" />,
     color: '#E34F26',
     name: 'HTML5',
   },
   instagram: {
-    svg: <InstagramSVG width="20px" height="20px" />,
+    svg: <InstagramSVG className="h-4 w-4" />,
     color: '#E4405F',
     name: 'Instagram',
     link: 'https://www.instagram.com/caffhheine/',
   },
   javascript: {
-    svg: <JavaScriptSVG width="20px" height="20px" />,
+    svg: <JavaScriptSVG className="h-4 w-4" />,
     color: '#F7DF1E',
     name: 'JavaScript',
   },
   naver: {
-    svg: <NaverSVG width="20px" height="20px" />,
+    svg: <NaverSVG className="h-4 w-4" />,
     color: '#03C75A',
     name: 'Naver',
     link: 'mailto:swarvy0826@naver.com',
   },
   nextjs: {
-    svg: <NextJsSVG width="20px" height="20px" />,
+    svg: <NextJsSVG className="h-4 w-4" />,
     color: '#000000',
     name: 'Next',
   },
   react: {
-    svg: <ReactSVG width="20px" height="20px" />,
+    svg: <ReactSVG className="h-4 w-4" />,
     color: '#61DAFB',
     name: 'React',
   },
   reactQuery: {
-    svg: <ReactQuerySVG width="20px" height="20px" />,
+    svg: <ReactQuerySVG className="h-4 w-4" />,
     color: '#FF4154',
     name: 'ReactQuery',
   },
   recoil: {
-    svg: <RecoilSVG width="20px" height="20px" />,
+    svg: <RecoilSVG className="h-4 w-4" />,
     color: '#3578E5',
     name: 'Recoil',
   },
   redux: {
-    svg: <ReduxSVG width="20px" height="20px" />,
+    svg: <ReduxSVG className="h-4 w-4" />,
     color: '#764ABC',
     name: 'Redux',
   },
   tailwindcss: {
-    svg: <TailwindcssSVG width="20px" height="20px" />,
+    svg: <TailwindcssSVG className="h-4 w-4" />,
     color: '#06B6D4',
     name: 'TailwindCSS',
   },
   typescript: {
-    svg: <TypeScriptSVG width="20px" height="20px" />,
+    svg: <TypeScriptSVG className="h-4 w-4" />,
     color: '#3178C6',
     name: 'TypeScript',
   },
   framer: {
-    svg: <FramerSVG width="20px" height="20px" />,
+    svg: <FramerSVG className="h-4 w-4" />,
     color: '#0055FF',
     name: 'Framer',
   },
   git: {
-    svg: <GitSVG width="20px" height="20px" />,
+    svg: <GitSVG className="h-4 w-4" />,
     color: '#F05032',
     name: 'Git',
   },
   axios: {
-    svg: <AxiosSVG width="20px" height="20px" />,
+    svg: <AxiosSVG className="h-4 w-4" />,
     color: '#5A29E4',
     name: 'Axios',
   },
   discord: {
-    svg: <DiscordSVG width="20px" height="20px" />,
+    svg: <DiscordSVG className="h-4 w-4" />,
     color: '#5865F2',
     name: 'Discord',
   },
